requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
python-dotenv
streamlit